move to a normal packaging structure
